Moved status-connection Service class into its own file. Also moved getApplicationStatus function into the StatusButton class, since that's where it's used.
